Fix some jumble code

+++ app/src/main/java/cc/flydev/launcher/ButtonDropTarget.java
@@ -24,6 +24,7 @@ import android.graphics.drawable.Drawable;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.TextView;
+
 import cc.flydev.face.R;
 
 
@@ -39,10 +40,14 @@ public class ButtonDropTarget extends TextView implements DropTarget, DragContro
     protected TextView mText;
     protected SearchDropTargetBar mSearchDropTargetBar;
 
-    /** Whether this drop target is active for the current drag */
+    /**
+     * Whether this drop target is active for the current drag
+     */
     protected boolean mActive;
 
-    /** The paint applied to the drag view on hover */
+    /**
+     * The paint applied to the drag view on hover
+     */
     protected int mHoverColor = 0;
 
     public ButtonDropTarget(Context context, AttributeSet attrs) {
@@ -146,13 +151,13 @@ public class ButtonDropTarget extends TextView implements DropTarget, DragContro
         }
 
         final int top = to.top + (getMeasuredHeight() - height) / 2;
-        final int bottom = top +  height;
+        final int bottom = top + height;
 
         to.set(left, top, right, bottom);
 
         // Center the destination rect about the trash icon
-        final int xOffset = (int) -(viewWidth - width) / 2;
-        final int yOffset = (int) -(viewHeight - height) / 2;
+        final int xOffset = -(viewWidth - width) / 2;
+        final int yOffset = -(viewHeight - height) / 2;
         to.offset(xOffset, yOffset);
 
         return to;
